socio-backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dj-rest-auth==7.0.1
Django==5.2.2
django-allauth==65.9.0
django-auth-adfs==1.15.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0

#Selenium dependencies
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
idna==3.10
lxml==5.4.0
outcome==1.3.0.post0
packaging==25.0
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
